refactor: delete not needed components
