+++ webclient/app/src/app/commons/confirmationDialog/ConfirmationDialog.jsx
