--- webclient/app/src/app/commons/form/AddUpdateFormStyles.js
@@ -1,9 +1,0 @@
-const AddUpdateFormStyles = theme => ({
-    textField: {
-        marginLeft: theme.spacing.unit,
-        marginRight: theme.spacing.unit,
-        width: 500
-    }
-});
-
-export default AddUpdateFormStyles;
